docs(pane): rename pane folder and story to paneprovider
