+++ packages/grid/src/elements/pane/components/Splitter.spec.tsx
