+++ packages/grid/src/elements/pane/components/Splitter.tsx
