+++ packages/grid/src/styled/pane/StyledPane.ts
