+++ packages/grid/src/styled/pane/StyledPaneContent.ts
