+++ packages/grid/src/styled/pane/StyledPaneSplitter.ts
